// libraries/googleapis/protos/google/type/money_pb.d.ts
// @generated by protoc-gen-es v2.4.0 with parameter "keep_empty_files=true,target=js+dts,import_extension=js"
// @generated from file libraries/googleapis/protos/google/type/money.proto (package google.type, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file libraries/googleapis/protos/google/type/money.proto.
 */
export declare const file_libraries_googleapis_protos_google_type_money: GenFile;

/**
 * @generated from message google.type.Money
 */
export declare type Money = Message<"google.type.Money"> & {
  /**
   * @generated from field: string currency_code = 1;
   */
  currencyCode: string;

  /**
   * @generated from field: int64 units = 2;
   */
  units: bigint;

  /**
   * @generated from field: int32 nanos = 3;
   */
  nanos: number;
};

/**
 * Describes the message google.type.Money.
 * Use `create(MoneySchema)` to create a new message.
 */
export declare const MoneySchema: GenMessage<Money>;